add httpurlconnection and gradle setting

+++ app/src/main/java/com/example/masstouring/mapactivity/MapActivity.java
@@ -155,15 +155,13 @@ public class MapActivity extends AppCompatActivity{
 
     @Override
     public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-        switch (item.getItemId()){
-            case R.id.action_delete:
-                DeleteConfirmationDialog dialog = new DeleteConfirmationDialog();
-                dialog.setCallback(oDeleteDialogCallback);
-                dialog.show(getSupportFragmentManager(), "deleteConfirmationDialog");
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
+        if (item.getItemId() == R.id.action_delete) {
+            DeleteConfirmationDialog dialog = new DeleteConfirmationDialog();
+            dialog.setCallback(oDeleteDialogCallback);
+            dialog.show(getSupportFragmentManager(), "deleteConfirmationDialog");
+            return true;
         }
+        return super.onOptionsItemSelected(item);
     }
 
     @Override
